+++ app/app/schemas/donate.py
@@ -1,5 +1,6 @@
 import uuid
 
+from decimal import Decimal
 from pydantic import BaseModel, Field
 
 from app.models.donate import DonateTransactionType
@@ -10,14 +11,14 @@ class DonateEntity(BaseModel):
     """Представление модели Donate"""
 
     telegram_user_id: uuid.UUID = Field(title="ID пользователя")
-    quantity: float = Field(title="Размер доната")
+    quantity: Decimal = Field(title="Размер доната")
     matrix_id: uuid.UUID = Field(title="ID матрицы")
 
 
 class BaseDonateTransactionSchema(BaseModel):
     sponsor_id: uuid.UUID = Field(title="ID спонсора")
     donate_id: uuid.UUID = Field(title="ID доната")
-    quantity: float = Field(title="Размер доната")
+    quantity: Decimal = Field(title="Размер доната")
     type_: DonateTransactionType
